etl-pipeline/requirements.txt
pdfplumber==0.11.4
elasticsearch==8.10.0
PyYAML==6.0.2
